Add required_fields for subclass templates; Add weather in the test file Restructure code base
