+++ rct229/rule_engine/rule_base.py
@@ -510,13 +510,14 @@ class RuleDefinitionListBase(RuleDefinitionBase):
     Baseclass for Rule Definitions that apply to each element in a list context.
     """
 
-    def __init__(self, rmrs_used, each_rule, id=None, description=None, rmr_context=""):
+    def __init__(self, rmrs_used, each_rule, id=None, description=None, rmr_context="", required_fields=None):
         self.each_rule = each_rule
         super(RuleDefinitionListBase, self).__init__(
             rmrs_used=rmrs_used,
             id=id,
             description=description,
             rmr_context=rmr_context,
+            required_fields=required_fields,
         )
 
     def create_context_list(self, context, data=None):
@@ -652,6 +653,7 @@ class RuleDefinitionListIndexedBase(RuleDefinitionListBase):
         rmr_context="",
         list_path="[*]",
         match_by="id",
+        required_fields = None,
     ):
         self.index_rmr = index_rmr
         self.list_path = list_path
@@ -662,6 +664,7 @@ class RuleDefinitionListIndexedBase(RuleDefinitionListBase):
             id=id,
             description=description,
             rmr_context=rmr_context,
+            required_fields=required_fields
         )
 
     def create_context_list(self, context, data=None):
