+++ rct229/rules/section5/section5rule5.py
@@ -1,0 +1,129 @@
+from rct229.data.schema_enums import schema_enums
+from rct229.data_fns.table_G3_4_fns import table_G34_lookup
+from rct229.rule_engine.rule_base import (
+    RuleDefinitionBase,
+    RuleDefinitionListIndexedBase,
+)
+from rct229.rule_engine.user_baseline_proposed_vals import UserBaselineProposedVals
+from rct229.ruleset_functions.get_opaque_surface_type import (
+    BELOW_GRADE_WALL,
+    ROOF,
+    get_opaque_surface_type,
+)
+from rct229.ruleset_functions.get_surface_conditioning_category_dict import (
+    EXTERIOR_MIXED,
+    EXTERIOR_NON_RESIDENTIAL,
+    EXTERIOR_RESIDENTIAL,
+    SEMI_EXTERIOR,
+    UNREGULATED,
+    get_surface_conditioning_category_dict,
+)
+from rct229.utils.jsonpath_utils import find_all
+
+
+class Section5Rule5(RuleDefinitionListIndexedBase):
+    """Rule 5 of ASHRAE 90.1-2019 Appendix G Section 5 (Envelope)"""
+
+    def __init__(self):
+        super(Section5Rule5, self).__init__(
+            rmrs_used=UserBaselineProposedVals(False, True, False),
+            required_fields={
+                "$": ["weather"],
+                "weather": ["climate_zone"],
+            },
+            each_rule=Section5Rule5.BuildingRule(),
+            index_rmr="baseline",
+            id="5-5",
+            description="Baseline roof assemblies must match the appropriate assembly maximum U-factors in Tables G3.4-1 through G3.4-8.",
+            list_path="buildings[*]",
+        )
+
+    def create_data(self, context, data=None):
+        rmr_baseline = context.baseline
+        return {"climate_zone": rmr_baseline["weather"]["climate_zone"]}
+
+    class BuildingRule(RuleDefinitionListIndexedBase):
+        def __init__(self):
+            super(Section5Rule5.BuildingRule, self).__init__(
+                rmrs_used=UserBaselineProposedVals(False, True, False),
+                required_fields={},
+                each_rule=Section5Rule5.BuildingRule.RoofRule(),
+                index_rmr="baseline",
+            )
+
+        def create_context_list(self, context, data=None):
+            building = context.baseline
+            # List of all baseline roof surfaces to become the context for RoofRule
+            return [
+                UserBaselineProposedVals(None, surface, None)
+                for surface in find_all("$..surfaces[*]", building)
+                if get_opaque_surface_type(surface) == ROOF
+            ]
+
+        def create_data(self, context, data=None):
+            building = context.baseline
+            # Merge into the existing data dict
+            return {
+                **data,
+                "surface_conditioning_category_dict": get_surface_conditioning_category_dict(
+                    data["climate_zone"], building
+                ),
+            }
+
+        class RoofRule(RuleDefinitionBase):
+            def __init__(self):
+                super(Section5Rule5.BuildingRule.RoofRule, self).__init__(
+                    rmrs_used=UserBaselineProposedVals(False, True, False),
+                    required_fields={},
+                )
+
+            def get_calc_vals(self, context, data=None):
+                climate_zone: str = data["climate_zone"]
+                roof = context.baseline
+                scc: str = data["surface_conditioning_category_dict"][roof["id"]]
+                roof_u_factor = roof["construction"]["u_factor"]
+
+                target_u_factor = None
+                target_u_factor_res = None
+                target_u_factor_nonres = None
+
+                if scc in [
+                    EXTERIOR_RESIDENTIAL,
+                    EXTERIOR_NON_RESIDENTIAL,
+                    SEMI_EXTERIOR,
+                ]:
+                    target_u_factor = table_G34_lookup(climate_zone, scc, ROOF)[
+                        "u_value"
+                    ]
+                elif scc == EXTERIOR_MIXED:
+                    target_u_factor_res = table_G34_lookup(
+                        climate_zone, EXTERIOR_RESIDENTIAL, ROOF
+                    )["u_value"]
+                    target_u_factor_nonres = table_G34_lookup(
+                        climate_zone, EXTERIOR_NON_RESIDENTIAL, ROOF
+                    )["u_value"]
+                    if target_u_factor_res == target_u_factor_nonres:
+                        target_u_factor = target_u_factor_res
+
+                return {
+                    "roof_u_factor": roof_u_factor,
+                    "target_u_factor": target_u_factor,
+                    "target_u_factor_res": target_u_factor_res,
+                    "target_u_factor_nonres": target_u_factor_nonres,
+                }
+
+            def manaul_check_required(self, context, calc_vals, data=None):
+                target_u_factor_res = calc_vals["target_u_factor_res"]
+                target_u_factor_nonres = calc_vals["target_u_factor_nonres"]
+
+                return (
+                    target_u_factor_res is not None
+                    and target_u_factor_nonres is not None
+                    and target_u_factor_res != target_u_factor_nonres
+                )
+
+            def rule_check(self, context, calc_vals, data=None):
+                roof_u_factor = calc_vals["roof_u_factor"]
+                target_u_factor = calc_vals["target_u_factor"]
+
+                return roof_u_factor == target_u_factor
