+++ test/test_section5.py
@@ -15,7 +15,7 @@ def test_rule5_8_success():
     proposed_dir = None
     user_dir = None
 
-    rule = rules.section5.Section5Rule8()
+    rule = rules.section5.section5rule8.Section5Rule8
     report = helper.evaluate_single_rule(user_dir, base_dir, proposed_dir, rule)
     results = helper.report_to_result_for_a_rule(report)
     assert results["result"] == "PASSED"
@@ -26,7 +26,7 @@ def test_rule5_8_one_surface_c_factor():
     proposed_dir = None
     user_dir = None
 
-    rule = rules.section5.Section5Rule8()
+    rule = rules.section5.section5rule8.Section5Rule8
     report = helper.evaluate_single_rule(user_dir, base_dir, proposed_dir, rule)
     results = helper.report_to_result_for_a_rule(report)
     assert (
